requirements.txt
maya
tweepy
Pillow